+++ server/tests/test_migrate_corpora_repo.py
@@ -3,7 +3,7 @@ import os.path
 import tempfile
 import unittest
 
-from clic.migrate.corpora_repo import import_book, export_book
+from clic.migrate.corpora_repo import parse_corpora_bib, import_book, export_book
 
 
 class RequiresCorporaDir():
@@ -93,3 +93,86 @@ class Test_export_book(unittest.TestCase, RequiresCorporaDir):
                 "chapter.text,0,12,,I'm a book. ",
                 "",
             ])
+
+
+class Test_parse_corpora_bib(unittest.TestCase, RequiresCorporaDir):
+    maxDiff = None
+
+    def test_call(self):
+        corpora_dir = self.corpora_dir({
+            'images/ChiLit_0.4.jpg': "JPEG!",
+            'corpora.bib': """
+@book{swift_gullivers_1726,
+        title = {Gulliver's Travels into Several Remote Nations of the World},
+        url = {https://www.gutenberg.org/ebooks/829},
+        shorttitle = {gulliver},
+        author = {Swift, Jonathan},
+        editor = {Price, David},
+        urldate = {2017-06-28},
+        date = {1726},
+        keywords = {{ArTs}}
+}
+
+@book{cermakova_childrens_2017,
+        location = {University of Birmingham, {UK}},
+        title = {Children's Literature},
+        series = {{CCR} Corpus},
+        shorttitle = {{ChiLit}},
+        number = {3},
+        publisher = {Centre for Corpus Research},
+        author = {Čermáková, A. and Mahlberg, M. and Wiegand, V.},
+        date = {2017},
+        keywords = {corpus}
+}
+
+@book{anstey_brass_1900,
+        title = {The Brass Bottle},
+        url = {https://www.gutenberg.org/ebooks/30689},
+        shorttitle = {brass},
+        author = {Anstey, F.},
+        urldate = {2017-09-10},
+        date = {1900},
+        keywords = {{ChiLit}}
+}
+
+@book{mahlberg_additional_2017,
+        location = {University of Birmingham, {UK}},
+        title = {Additional Requested Texts},
+        series = {{CCR} Corpus},
+        shorttitle = {{ArTs}},
+        number = {4},
+        publisher = {Centre for Corpus Research},
+        author = {Mahlberg, M. and Wiegand, V. and Čermáková, A.},
+        date = {2017},
+        keywords = {corpus}
+}
+
+@book{crockett_surprising_1897,
+        title = {The Surprising Adventures of Sir Toady Lion with Those of General Napoleon Smith},
+        url = {https://www.gutenberg.org/ebooks/39340},
+        shorttitle = {toadylion},
+        author = {Crockett, S. R.},
+        urldate = {2017-09-10},
+        date = {1897},
+        keywords = {{ArTs},{ChiLit}}
+}
+            """.strip(),
+        })
+        out = sorted(parse_corpora_bib(corpora_dir), key=lambda c: c['ordering'])
+        self.assertEqual(out, [
+            dict(
+                name='ChiLit',
+                title="Children's Literature",
+                description='',
+                ordering=3,
+                carousel_image_path=os.path.join(corpora_dir, 'images', 'ChiLit_0.4.jpg'),
+                contents=['brass', 'toadylion'],
+            ), dict(
+                name='ArTs',
+                title='Additional Requested Texts',
+                description='',
+                ordering=4,
+                carousel_image_path=None,
+                contents=['gulliver', 'toadylion'],
+            ),
+        ])